refactor: split food recipes app
